mass generating blue orbs progress

+++ src/components/GrayRing.tsx
@@ -62,8 +62,9 @@ const GrayRing = memo((props: GrayRingProps) => {
     vec4 color(vec2 vUv, int quadnum, bool textureexists, int thinperiod, int quadlen, float step) {
       if (!textureexists) {
         return vec4(0.259,0.259,0.322, 1);
-      } else if (quadnum % 2 == 1) {
+      } else if (uint(quadnum) % uint(2) == uint(1)) {
         return texture2D(hole, vec2(tolocal(vUv.x, quadlen-thinperiod, step), vUv.y));
+          // return vec4(tolocal(vUv.x, quadlen-thinperiod, step), 0, 0, 1);
       } else if (quadnum == 0) {
         return texture2D(lof, vec2(tolocal(vUv.x, quadlen-thinperiod, step), vUv.y));
       } else {
@@ -84,18 +85,18 @@ const GrayRing = memo((props: GrayRingProps) => {
       int quadlen = int(step)/4;
       
       // segment within circle's quad
-      int quadel = int(segment) % quadlen;
+      uint quadel = uint(segment) % uint(quadlen);
       
       // which quad
-      int quadnum = int(segment) / quadlen;
+      int quadnum = int(uint(segment) / uint(quadlen));
      
      // how big thin part is
      int thinperiod = 8;
      
-      if (quadel < thinperiod && isheight(vUv.y, thin)) {
+      if (quadel < uint(thinperiod) && isheight(vUv.y, thin)) {
           // thin line
           gl_FragColor = color(vUv, quadnum, false, thinperiod, quadlen, step);
-      } else if (quadel == thinperiod) {
+      } else if (quadel == uint(thinperiod)) {
           // slope up
           float dist = tolocal(vUv.x, 1, step);
           if (vUv.y > slope(1.0-dist, thin) && vUv.y < 1.0-slope(1.0-dist, thin)) {
@@ -103,7 +104,7 @@ const GrayRing = memo((props: GrayRingProps) => {
           } else {
               gl_FragColor = vec4(0, 0, 0, 0);
           }
-      } else if (quadel == quadlen-1) {
+      } else if (quadel == uint(quadlen-1)) {
           // slope down
           float dist = tolocal(vUv.x, 1, step);
           if (vUv.y > slope(dist, thin) && vUv.y < 1.0-slope(dist, thin)) {
@@ -111,7 +112,7 @@ const GrayRing = memo((props: GrayRingProps) => {
           } else {
               gl_FragColor = vec4(0, 0, 0, 0);
           }
-      } else if (quadel > thinperiod) {
+      } else if (quadel > uint(thinperiod)) {
             gl_FragColor = color(vUv, quadnum, true, thinperiod, quadlen, step);
       } else {
           // transparent
